ANDROID: ASan abort code change
SIGABRT is not a monitored signal (thus 'abort_on_error' is missing crashes when set)
for Android OS since it produces lots of useless crashes due to way Android process
termination hacks work. Safest option is to register & monitor one of user signals.
SIGUSR2 is used for sanitizer fuzzing in Android, although might need to be changed
if target uses it for other purposes.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>