SANCOV: Don't mangle 1st iteration
In order to have accurate coverage data to work against
the first iteration of a new seed pickup is not mangled. This
will save the coverage bitmaps of original input.

In case of multiple worker threads, only one picks this tasks
and keeps a lock until finished, blocking other threads from
continuing fuzzing.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>