LINUX: Add read PID from file
Similar to '-p' flag (PID provided as a number), '-P' flag
is added to read PID from file. Many deamonized targets
offer the option to write running PID to file. We can now
read it directly and benefit from cases of auto-restart so
we don't need to launch again the campaign when remote
target crashed.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>