[android] Fix wrong linking with libhfuzz.a from main tool
libhfuzz is not an actual dependency of the honggfuzz standalone tool. Instead it
is simply desired to always build it with the build config matching the main tool so
that it is available to user for wrapping targets with instrumentation.

Having libhfuzz as part of LOCAL_STATIC_LIBRARIES effectively instructs NDK
automation to include it as dependency. This results into resolving some of the
external libc dependencies (strcmp, strstr, etc.) with the ones implemented from
libhfuzz. This is not desired since it breaks functionality.

Replace LOCAL_STATIC_LIBRARIES dependency with a proper APP_MODULES
list when invoking ndk-build to always build both modules.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>