add new statistic collection
add crashes count and timeouted count to the final summary
add new_units_added libfuzzer like statitic - counts the number of new samples added in the corpus during this run because of new coverage found.
add slowest unit - after each round, update the timeOfLongestUnitInMilliseconds if the round was more than 10% slower than the slowest one so far and above option threshold (default: 10s). Timeouts are not considered.
print the statistics in the final summary. The slowest unit value will always be 0 if under the threshold defined in the options.