netdriver: make it possible to specify both type (e.g. SOCK_DGRAM) and protocol for the destination address.
This requires change to the HonggfuzzNetDriverServerAddress()
declaration, as it needs to enable its users to specify those two
additional params.

Additionally, pass struct sockaddr_storage to
HonggfuzzNetDriverServerAddress(), as this will prevent this func for
owning storage for the used sockaddr. It was not a bug, but this
semantics should be more familiar for C coders.