cmdline: fail if command line can't be parsed
It should make it easier to integrate honggfuzz into CI.

Trying to integrate honggfuzz into the elfutils testsuite I
wrote a script running fuzz targets with hongfuzz for a couple
of minutes. It worked with honggfuzz built using the master
branch but when I switched to the latest stable version a test
that should have failed passed unexpectedly. The culprit was
--exit_code_upon_crash (which hasn't been released yet). If would
have been much easier to track that down if honggfuzz had complained
that it failed to parse the command line and exited with a non-zero
exit code.